20240910 commit 20240812 adjusted GWF so that minimum f2 from smooth and rough conditions is used for kn>0 to prevent unrealistically high influence pres-gradient for low dpdl and rough conditions (reasoning being that for rough conditions f2 needs to be lower than hydr-smooth conditions because higher kn lowers f2, so rough f2 should never be larger than smooth f2) fix small typo in determination TBLEsed_dpdx,TBLEsed_dpdy for wm8 and wm9 fix incorrectly making dpdx=0. and dpdy=0. for wm9 for all cells and moments with adverse pressure gradient; this made relaxated pressure gradient incorrect fix incorrect use of variables rhoU,rhoV,rhoW in fkdat even when they were not declared fixed some coding issues after stringent debug compilation test run fixed important error in vfftf.f fft-solving-routines; for vradbg and vradfg (general kernel when 4,2,3,5 kernels cannot be used). An important help array was defined WA(IDO), but this should be WA(IDO*(IP-1)), otherwise out-of-bound errors and unexpected failure for some simulations with unfortunate kmax which ended up needing vradbg/vradfg with IP>1.

+++ rheology.f
@@ -23,7 +23,7 @@
 	!include 'mpif.h'
 	
 	integer ib,ie,jb,je,kb,ke
-	integer kbed_dummy(0:i1,0:j1)
+	integer*8 kbed_dummy(0:i1,0:j1)
 	real lambda_advec(0:i1,0:j1,0:k1)
 	
 	lambda_advec(:,:,:)=0.
